Split UP adapters into v2 and v3
